+++ src/applications/merchant-app/components/CustomColumn.svelte
@@ -7,17 +7,34 @@
 
 	const doc = item.itemDocument;
 
-	let text;
+	let text = "";
+	let value = "";
+	let title = ""
 
-	$: value = data.path ? foundry.utils.getProperty($doc, data.path) ?? "" : "";
 	$: {
-		let localized = localize(`${data.mapping?.[value] ?? value}`);
-		text = data.formatting ? data.formatting.replace("{#}", localized) : localized;
+		if (data.condition && foundry.utils.getProperty($doc, data.condition.path) === data.condition.value) {
+			value = data.condition.placeholder ?? "";
+			text = data.condition.placeholder ? localize(value) : "";
+
+			if (game.user.isGM) {
+				let hidden_value = data.path ? foundry.utils.getProperty($doc, data.path) ?? "" : "";
+				let localized = localize(`${data.mapping?.[hidden_value] ?? hidden_value}`);
+				title = data.formatting ? data.formatting.replace("{#}", localized) : localized;
+				if (hidden_value) {
+					text += "*"
+				}
+			}
+		} else {
+			value = data.path ? foundry.utils.getProperty($doc, data.path) ?? "" : "";
+			let localized = localize(`${data.mapping?.[value] ?? value}`);
+			text = data.formatting ? data.formatting.replace("{#}", localized) : localized;
+			title = "";
+		}
 	}
 
 </script>
 
-<div>
+<div title="{title}">
 	<span>{@html text}</span>
 </div>